Updated Monitor class and added file events
